refactor: move impact metrics to a separate page
