+++ frontend/src/component/impact-metrics/hooks/useSeriesColor.ts
