+++ frontend/src/component/impact-metrics/utils.ts
